refactor: extract MCP SSE routes from http-server into dedicated route module
Move ~530 lines of MCP SSE endpoints (OAuth discovery, SSE connections,
v1/sse standard MCP transport, /mcp discovery, OAuth redirects) into
routes/mcp-sse-routes.ts. http-server.ts reduced from 1,912 to 1,338 lines.
