+++ mcp_backend/src/http-server.ts
@@ -15,6 +15,7 @@ import { createRestAPIRouter } from './routes/rest-api.js';
 import { createBillingInlineRoutes } from './routes/billing-inline-routes.js';
 import { createChatInlineRoutes } from './routes/chat-inline-routes.js';
 import { createMiscInlineRoutes } from './routes/misc-inline-routes.js';
+import { createMCPSSERoutes } from './routes/mcp-sse-routes.js';
 import { createBalanceCheckMiddleware } from './middleware/balance-check.js';
 import { createPaymentRouter, createWebhookRouter } from './routes/payment-routes.js';
 import { createBillingRoutes } from './routes/billing-routes.js';
@@ -37,12 +38,10 @@ import { getRedisClient } from './utils/redis-client.js';
 import { CacheAdapter } from './infrastructure/adapters/cache-adapter.js';
 import { LLMAdapter } from './infrastructure/adapters/llm-adapter.js';
 import { createTemplateRoutes } from './routes/template-routes.js';
-import { Server } from '@modelcontextprotocol/sdk/server/index.js';
 import { SSEServerTransport } from '@modelcontextprotocol/sdk/server/sse.js';
-import { CallToolRequestSchema, ListToolsRequestSchema } from '@modelcontextprotocol/sdk/types.js';
 import { createOAuthRouter } from './routes/oauth-routes.js';
 // createHybridAuthMiddleware available from './middleware/oauth-auth.js' if needed
-import { mcpDiscoveryRateLimit, healthCheckRateLimit, webhookRateLimit, chatRateLimit, globalApiRateLimit } from './middleware/rate-limit.js';
+import { healthCheckRateLimit, webhookRateLimit, globalApiRateLimit } from './middleware/rate-limit.js';
 import { ServiceType } from './types/gateway.js';
 import { createUploadRouter } from './routes/upload-routes.js';
 import { createConversationRouter } from './routes/conversation-routes.js';
@@ -65,7 +64,6 @@ import { createJudgesRoutes } from './routes/judges-routes.js';
 import { JudgeAnalyticsService } from './services/judge-analytics-service.js';
 import { createJudgeAnalyticsRoutes } from './routes/judge-analytics-routes.js';
 import { createReferralRoutes } from './routes/referral-routes.js';
-import { sanitizeId, maskSensitive } from './utils/sanitize-log.js';
 import rateLimit from 'express-rate-limit';
 
 dotenv.config();
@@ -682,591 +680,19 @@ class HTTPMCPServer {
       });
     });
 
-    // OPTIONS handler for /sse - returns OAuth configuration
-    // This allows ChatGPT to discover OAuth endpoints
-    this.app.options('/sse', (req: Request, res: Response) => {
-      const proto = req.headers['x-forwarded-proto'] || req.protocol || 'https';
-      const host = req.headers['x-forwarded-host'] || req.headers.host;
-      const baseUrl = process.env.PUBLIC_URL || `${proto}://${host}`;
-
-      res.setHeader('MCP-Auth-Type', 'oauth2');
-      res.setHeader('MCP-Auth-Authorization-Endpoint', `${baseUrl}/oauth/authorize`);
-      res.setHeader('MCP-Auth-Token-Endpoint', `${baseUrl}/oauth/token`);
-      res.setHeader('MCP-Auth-Scopes', 'mcp');
-      res.setHeader('Allow', 'GET, POST, OPTIONS');
-      res.status(200).send();
-    });
-
-    // GET handler for /sse - returns OAuth configuration as JSON
-    // ChatGPT may use this for discovery
-    this.app.get('/sse', (req: Request, res: Response) => {
-      const proto = req.headers['x-forwarded-proto'] || req.protocol || 'https';
-      const host = req.headers['x-forwarded-host'] || req.headers.host;
-      const baseUrl = process.env.PUBLIC_URL || `${proto}://${host}`;
-
-      res.json({
-        protocol: 'mcp',
-        version: '1.0',
-        auth: {
-          type: 'oauth2',
-          authorization_endpoint: `${baseUrl}/oauth/authorize`,
-          token_endpoint: `${baseUrl}/oauth/token`,
-          scopes: ['mcp'],
-        },
-        capabilities: {
-          tools: true,
-        },
-      });
-    });
-
-    // OAuth 2.0 Authorization Server Metadata (RFC 8414)
-    // ChatGPT checks this for OAuth discovery
-    this.app.get('/sse/.well-known/oauth-authorization-server', (req: Request, res: Response) => {
-      const proto = req.headers['x-forwarded-proto'] || req.protocol || 'https';
-      const host = req.headers['x-forwarded-host'] || req.headers.host;
-      const baseUrl = process.env.PUBLIC_URL || `${proto}://${host}`;
-
-      res.json({
-        issuer: baseUrl,
-        authorization_endpoint: `${baseUrl}/oauth/authorize`,
-        token_endpoint: `${baseUrl}/oauth/token`,
-        registration_endpoint: `${baseUrl}/oauth/register`,
-        revocation_endpoint: `${baseUrl}/oauth/revoke`,
-        response_types_supported: ['code'],
-        grant_types_supported: ['authorization_code'],
-        token_endpoint_auth_methods_supported: ['none', 'client_secret_post'],
-        scopes_supported: ['mcp', 'claudeai'],
-        code_challenge_methods_supported: ['S256', 'plain'],
-      });
-    });
-
-    // OpenID Connect Discovery (for compatibility)
-    this.app.get('/sse/.well-known/openid-configuration', (req: Request, res: Response) => {
-      const proto = req.headers['x-forwarded-proto'] || req.protocol || 'https';
-      const host = req.headers['x-forwarded-host'] || req.headers.host;
-      const baseUrl = process.env.PUBLIC_URL || `${proto}://${host}`;
-
-      res.json({
-        issuer: baseUrl,
-        authorization_endpoint: `${baseUrl}/oauth/authorize`,
-        token_endpoint: `${baseUrl}/oauth/token`,
-        registration_endpoint: `${baseUrl}/oauth/register`,
-        revocation_endpoint: `${baseUrl}/oauth/revoke`,
-        response_types_supported: ['code'],
-        grant_types_supported: ['authorization_code'],
-        token_endpoint_auth_methods_supported: ['none', 'client_secret_post'],
-        scopes_supported: ['mcp', 'claudeai'],
-        code_challenge_methods_supported: ['S256', 'plain'],
-      });
-    });
-
-    // MCP SSE endpoint for ChatGPT web integration (REQUIRED auth)
-    // Endpoint: POST /sse
-    // This implements the Model Context Protocol over Server-Sent Events
-    // Reference: https://platform.openai.com/docs/mcp
-    this.app.post('/sse', (async (req: DualAuthRequest, res: Response) => {
-      try {
-        // CRITICAL: Authentication is REQUIRED for usage tracking
-        const authHeader = req.headers.authorization;
-        if (!authHeader || !authHeader.startsWith('Bearer ')) {
-          logger.warn('[MCP SSE] Missing or invalid Authorization header');
-          return res.status(401).json({
-            error: 'Unauthorized',
-            message: 'Authorization header with Bearer token is required',
-            code: 'MISSING_AUTH',
-          });
-        }
-
-        const token = authHeader.replace('Bearer ', '');
-        let userId: string | undefined;
-        let clientKey: string | undefined;
-
-        // Authenticate (JWT, OAuth, or API key)
-        try {
-          if (token.includes('.')) {
-            // JWT token - verify and extract userId
-            const jwt = await import('jsonwebtoken');
-            const decoded = jwt.verify(token, process.env.JWT_SECRET || 'change-this-secret-in-production') as any;
-            userId = decoded.userId;
-            logger.debug('[MCP SSE] Authenticated with JWT', { userId });
-          } else if (token.startsWith('mcp_token_')) {
-            // OAuth 2.0 access token - verify with OAuth service
-            const tokenData = await this.app_.oauthService.verifyAccessToken(token);
-
-            if (!tokenData) {
-              logger.warn('[MCP SSE] Invalid OAuth token', {
-                tokenPrefix: maskSensitive(token, 15),
-              });
-              return res.status(401).json({
-                error: 'Unauthorized',
-                message: 'Invalid or expired OAuth access token',
-                code: 'INVALID_OAUTH_TOKEN',
-              });
-            }
-
-            userId = tokenData.userId;
-            clientKey = tokenData.clientId;
-            logger.debug('[MCP SSE] Authenticated with OAuth token', {
-              userId,
-              clientId: tokenData.clientId,
-              scope: tokenData.scope,
-            });
-          } else {
-            // API key - validate and get user info
-            clientKey = token;
-            const keyInfo = await this.billing.apiKeyService.validateApiKey(token);
-
-            if (!keyInfo) {
-              logger.warn('[MCP SSE] Invalid API key', {
-                keyPrefix: maskSensitive(token, 12),
-              });
-              return res.status(401).json({
-                error: 'Unauthorized',
-                message: 'Invalid API key',
-                code: 'INVALID_API_KEY',
-              });
-            }
-
-            // Valid API key - check rate limits
-            const rateLimit = await this.billing.apiKeyService.checkRateLimit(token);
-
-            if (!rateLimit.allowed) {
-              logger.warn('[MCP SSE] Rate limit exceeded', {
-                keyId: keyInfo.id,
-                reason: rateLimit.reason,
-              });
-              return res.status(429).json({
-                error: 'Rate limit exceeded',
-                code: 'RATE_LIMIT_EXCEEDED',
-                reason: rateLimit.reason,
-                requestsToday: rateLimit.requestsToday,
-                rateLimitPerDay: rateLimit.rateLimitPerDay,
-              });
-            }
-
-            // Get userId from API key
-            userId = keyInfo.userId;
-            logger.debug('[MCP SSE] Authenticated with API key', {
-              userId,
-              keyId: keyInfo.id,
-              userEmail: keyInfo.userEmail,
-            });
-
-            // Update API key usage (async, don't wait)
-            this.billing.apiKeyService.updateUsage(token).catch((err) => {
-              logger.error('[MCP SSE] Failed to update API key usage', { error: err.message });
-            });
-          }
-        } catch (error) {
-          // Auth failed - return 401
-          logger.warn('[MCP SSE] Authentication failed', { error: (error as Error).message });
-          return res.status(401).json({
-            error: 'Unauthorized',
-            message: 'Authentication failed: ' + (error as Error).message,
-            code: 'AUTH_FAILED',
-          });
-        }
-
-        // Authentication successful - set SSE headers and handle connection
-        res.setHeader('Content-Type', 'text/event-stream');
-        res.setHeader('Cache-Control', 'no-cache, no-transform');
-        res.setHeader('Connection', 'keep-alive');
-        res.setHeader('X-Accel-Buffering', 'no');
-
-        // Pass userId and clientKey to SSE handler
-        await this.app_.mcpSSEServer.handleSSEConnection(req, res, userId, clientKey);
-      } catch (error: any) {
-        logger.error('[MCP SSE] Connection error:', error);
-        if (!res.headersSent) {
-          res.status(500).json({
-            error: 'MCP SSE connection failed',
-            message: error.message,
-          });
-        }
-      }
-    }) as any);
-
-    // Standard MCP SSE endpoint for MCP clients (Claude Desktop, Jan chat, etc.)
-    // Endpoint: GET/POST /v1/sse (SSE stream + client messages)
-    // This implements the standard Model Context Protocol over SSE Transport
-    // Reference: https://spec.modelcontextprotocol.io/specification/transports/#server-sent-events
-
-    // POST handler: route messages to existing SSE sessions
-    this.app.post('/v1/sse', (async (req: DualAuthRequest, res: Response) => {
-      const sessionId = req.query.sessionId as string;
-      if (!sessionId) {
-        return res.status(400).json({ error: 'Missing sessionId query parameter' });
-      }
-
-      const transport = this.mcpSseSessions.get(sessionId);
-      if (!transport) {
-        return res.status(404).json({ error: 'Session not found', sessionId });
-      }
-
-      await transport.handlePostMessage(req, res, req.body);
-    }) as any);
-
-    // GET handler: establish new SSE stream
-    this.app.get('/v1/sse', (async (req: DualAuthRequest, res: Response) => {
-      try {
-        logger.info('[MCP v1/sse] New standard MCP SSE connection');
-
-        // REQUIRED authentication for usage tracking
-        const authHeader = req.headers.authorization;
-        if (!authHeader || !authHeader.startsWith('Bearer ')) {
-          logger.warn('[MCP v1/sse] Missing or invalid Authorization header');
-          return res.status(401).json({
-            error: 'Unauthorized',
-            message: 'Authorization header with Bearer token is required',
-            code: 'MISSING_AUTH',
-          });
-        }
-
-        const token = authHeader.replace('Bearer ', '');
-        let userId: string | undefined;
-        let clientKey: string | undefined;
-
-        // Authenticate (JWT, OAuth access token, or API key)
-        try {
-          if (token.includes('.')) {
-            // JWT token
-            const jwt = await import('jsonwebtoken');
-            const decoded = jwt.verify(token, process.env.JWT_SECRET || 'change-this-secret-in-production') as any;
-            userId = decoded.userId;
-            logger.debug('[MCP v1/sse] Authenticated with JWT', { userId });
-          } else {
-            // Try OAuth access token first, then API key
-            const oauthToken = await this.app_.oauthService.verifyAccessToken(token);
-            if (oauthToken) {
-              userId = oauthToken.userId;
-              logger.debug('[MCP v1/sse] Authenticated with OAuth token', { userId: sanitizeId(userId || ''), clientId: sanitizeId(oauthToken.clientId) });
-            } else {
-              // API key
-              clientKey = token;
-              const keyInfo = await this.billing.apiKeyService.validateApiKey(token);
-
-              if (!keyInfo) {
-                logger.warn('[MCP v1/sse] Invalid API key', {
-                  keyPrefix: maskSensitive(token, 8),
-                });
-                return res.status(401).json({
-                  error: 'Unauthorized',
-                  message: 'Invalid API key',
-                  code: 'INVALID_API_KEY',
-                });
-              }
-
-              // Check rate limits
-              const rateLimit = await this.billing.apiKeyService.checkRateLimit(token);
-
-              if (!rateLimit.allowed) {
-                logger.warn('[MCP v1/sse] Rate limit exceeded', {
-                  keyId: keyInfo.id,
-                  reason: rateLimit.reason,
-                });
-                return res.status(429).json({
-                  error: 'Rate limit exceeded',
-                  code: 'RATE_LIMIT_EXCEEDED',
-                  reason: rateLimit.reason,
-                });
-              }
-
-              userId = keyInfo.userId;
-              logger.debug('[MCP v1/sse] Authenticated with API key', {
-                userId,
-                keyId: keyInfo.id,
-              });
-
-              // Update API key usage
-              this.billing.apiKeyService.updateUsage(token).catch((err) => {
-                logger.error('[MCP v1/sse] Failed to update API key usage', { error: err.message });
-              });
-            }
-          }
-        } catch (error) {
-          // Auth failed - return 401
-          logger.warn('[MCP v1/sse] Authentication failed', { error: (error as Error).message });
-          return res.status(401).json({
-            error: 'Unauthorized',
-            message: 'Authentication failed: ' + (error as Error).message,
-            code: 'AUTH_FAILED',
-          });
-        }
-
-        // Create MCP Server instance for this connection
-        const mcpServer = new Server(
-          {
-            name: 'secondlayer-mcp',
-            version: '1.0.0',
-          },
-          {
-            capabilities: {
-              tools: {},
-            },
-          }
-        );
-
-        // Setup tools/list handler
-        mcpServer.setRequestHandler(ListToolsRequestSchema, async () => {
-          return {
-            tools: this.tools.toolRegistry.getLocalToolDefinitions(),
-          };
-        });
-
-        // Setup tools/call handler with billing integration
-        mcpServer.setRequestHandler(CallToolRequestSchema, async (request) => {
-          const toolName = request.params.name;
-          const args = request.params.arguments || {};
-          const requestId = `mcp-v1-${uuidv4()}`;
-          const startTime = Date.now();
-
-          try {
-            logger.info('[MCP v1/sse] Tool call', {
-              tool: toolName,
-              userId: sanitizeId(userId || 'anonymous'),
-            });
-
-            // Phase 2 Billing: Check credits BEFORE execution
-            if (userId && this.billing.creditService) {
-              const creditsRequired = await this.billing.creditService.calculateCreditsForTool(toolName, userId);
-
-              if (creditsRequired > 0) {
-                const balance = await this.billing.creditService.checkBalance(userId, creditsRequired);
-
-                if (!balance.hasCredits) {
-                  logger.warn('[MCP v1/sse] Insufficient credits', {
-                    userId: sanitizeId(userId),
-                    tool: toolName,
-                    creditsRequired,
-                  });
-
-                  return {
-                    content: [
-                      {
-                        type: 'text',
-                        text: `Error: Insufficient credits. Required: ${creditsRequired}, Current balance: ${balance.currentBalance}`,
-                      },
-                    ],
-                    isError: true,
-                  };
-                }
-              }
-            }
-
-            // Create cost tracking record
-            await this.billing.costTracker.createTrackingRecord({
-              requestId,
-              toolName,
-              clientKey,
-              userId,
-              userQuery: String(args.query || JSON.stringify(args)),
-              queryParams: args,
-            });
-
-            // Execute tool in request context
-            const result = await requestContext.run(
-              { requestId, task: toolName },
-              async () => {
-                // Route to appropriate tool handler via centralized registry
-                // Inject userId for all vault tools (user isolation)
-                const VAULT_TOOLS = new Set(['store_document', 'get_document', 'list_documents', 'semantic_search', 'list_folders', 'delete_document', 'update_document']);
-                const toolArgs = VAULT_TOOLS.has(toolName) ? { ...args, userId } : args;
-                const registryResult = await this.tools.toolRegistry.executeTool(toolName, toolArgs);
-                if (registryResult) {
-                  return registryResult;
-                }
-                throw new Error(`Unknown tool: ${toolName}`);
-              }
-            );
-
-            // Complete cost tracking
-            const executionTime = Date.now() - startTime;
-            await this.billing.costTracker.completeTrackingRecord({
-              requestId,
-              executionTimeMs: executionTime,
-              status: 'completed',
-            });
-
-            // Phase 2 Billing: Deduct credits after successful execution
-            if (userId && this.billing.creditService) {
-              const creditsRequired = await this.billing.creditService.calculateCreditsForTool(toolName, userId);
-
-              if (creditsRequired > 0) {
-                const deduction = await this.billing.creditService.deductCredits(
-                  userId,
-                  creditsRequired,
-                  toolName,
-                  requestId,
-                  `Tool execution: ${toolName}`
-                );
-
-                if (deduction.success) {
-                  logger.info('[MCP v1/sse] Credits deducted', {
-                    userId: sanitizeId(userId),
-                    tool: toolName,
-                    creditsDeducted: creditsRequired,
-                    newBalance: deduction.newBalance,
-                  });
-                }
-              }
-            }
-
-            // Return result in MCP format
-            return {
-              content: result.content || [
-                {
-                  type: 'text',
-                  text: JSON.stringify(result, null, 2),
-                },
-              ],
-            };
-
-          } catch (error: any) {
-            logger.error('[MCP v1/sse] Tool execution error', {
-              tool: toolName,
-              error: error.message,
-            });
-
-            // Record failure
-            const executionTime = Date.now() - startTime;
-            await this.billing.costTracker.completeTrackingRecord({
-              requestId,
-              executionTimeMs: executionTime,
-              status: 'failed',
-              errorMessage: error.message,
-            });
-
-            return {
-              content: [
-                {
-                  type: 'text',
-                  text: `Error: ${error.message}`,
-                },
-              ],
-              isError: true,
-            };
-          }
-        });
-
-        // Create SSE transport
-        const transport = new SSEServerTransport('/v1/sse', res);
-
-        // Store session for POST message routing
-        this.mcpSseSessions.set(transport.sessionId, transport);
-
-        // Connect MCP server to transport
-        await mcpServer.connect(transport);
-
-        logger.info('[MCP v1/sse] Connection established', { sessionId: transport.sessionId });
-
-        // Handle client disconnect
-        req.on('close', () => {
-          logger.info('[MCP v1/sse] Client disconnected', { sessionId: transport.sessionId });
-          this.mcpSseSessions.delete(transport.sessionId);
-          mcpServer.close();
-        });
-
-      } catch (error: any) {
-        logger.error('[MCP v1/sse] Connection error:', error);
-        if (!res.headersSent) {
-          res.status(500).json({
-            error: 'Failed to establish MCP SSE connection',
-            message: error.message,
-          });
-        }
-      }
-    }) as any);
-
-    // MCP discovery endpoint (public - lists available tools, rate limited)
-    // GET /mcp - Returns MCP server info and capabilities
-    this.app.get('/mcp', mcpDiscoveryRateLimit as any, (_req: Request, res: Response) => {
-      const tools = this.app_.mcpSSEServer.getAllTools();
-      res.json({
-        protocolVersion: '2024-11-05',
-        serverInfo: {
-          name: 'SecondLayer Legal MCP Server',
-          version: '1.0.0',
-          description: 'Ukrainian legal research and document analysis platform',
-        },
-        capabilities: {
-          tools: {
-            count: tools.length,
-            listChanged: false,
-          },
-          prompts: {},
-          resources: {},
-        },
-        endpoints: {
-          sse: '/sse',
-          'sse-standard': '/v1/sse',
-          http: '/api/tools',
-        },
-        tools: tools.map(t => ({
-          name: t.name,
-          description: t.description,
-        })),
-      });
-    });
+    // MCP SSE routes (SSE endpoints, OAuth discovery, redirects)
+    this.app.use(createMCPSSERoutes({
+      mcpSSEServer: this.app_.mcpSSEServer,
+      toolRegistry: this.tools.toolRegistry,
+      oauthService: this.app_.oauthService,
+      apiKeyService: this.billing.apiKeyService,
+      costTracker: this.billing.costTracker,
+      creditService: this.billing.creditService,
+      mcpSseSessions: this.mcpSseSessions,
+    }));
 
     // Authentication routes (public - OAuth endpoints, optional JWT for /auth/me etc.)
     this.app.use('/auth', optionalJWT as any, authRouter);
-
-    // Redirect /authorize to /oauth/authorize (for Claude.ai compatibility)
-    this.app.get('/authorize', (req: Request, res: Response) => {
-      const queryString = new URLSearchParams(req.query as any).toString();
-      res.redirect(301, `/oauth/authorize?${queryString}`);
-    });
-
-    this.app.post('/authorize', (req: Request, res: Response) => {
-      res.redirect(307, '/oauth/authorize');
-    });
-
-    // Redirect /token to /oauth/token (for Claude.ai compatibility)
-    this.app.post('/token', (req: Request, res: Response) => {
-      res.redirect(307, '/oauth/token');
-    });
-
-    // Root-level .well-known endpoints for OAuth discovery (Claude.ai compatibility)
-    this.app.get('/.well-known/oauth-authorization-server', (req: Request, res: Response) => {
-      const proto = req.headers['x-forwarded-proto'] || req.protocol || 'https';
-      const host = req.headers['x-forwarded-host'] || req.headers.host;
-      const baseUrl = process.env.PUBLIC_URL || `${proto}://${host}`;
-      res.json({
-        issuer: baseUrl,
-        authorization_endpoint: `${baseUrl}/oauth/authorize`,
-        token_endpoint: `${baseUrl}/oauth/token`,
-        registration_endpoint: `${baseUrl}/oauth/register`,
-        revocation_endpoint: `${baseUrl}/oauth/revoke`,
-        response_types_supported: ['code'],
-        grant_types_supported: ['authorization_code'],
-        token_endpoint_auth_methods_supported: ['none', 'client_secret_post'],
-        scopes_supported: ['mcp', 'claudeai'],
-        code_challenge_methods_supported: ['S256', 'plain'],
-      });
-    });
-
-    this.app.get('/.well-known/openid-configuration', (req: Request, res: Response) => {
-      const proto = req.headers['x-forwarded-proto'] || req.protocol || 'https';
-      const host = req.headers['x-forwarded-host'] || req.headers.host;
-      const baseUrl = process.env.PUBLIC_URL || `${proto}://${host}`;
-      res.json({
-        issuer: baseUrl,
-        authorization_endpoint: `${baseUrl}/oauth/authorize`,
-        token_endpoint: `${baseUrl}/oauth/token`,
-        registration_endpoint: `${baseUrl}/oauth/register`,
-        revocation_endpoint: `${baseUrl}/oauth/revoke`,
-        response_types_supported: ['code'],
-        grant_types_supported: ['authorization_code'],
-        token_endpoint_auth_methods_supported: ['none', 'client_secret_post'],
-        scopes_supported: ['mcp', 'claudeai'],
-        code_challenge_methods_supported: ['S256', 'plain'],
-      });
-    });
-
-    // Redirect /register to /oauth/register (for MCP client compatibility)
-    this.app.post('/register', (req: Request, res: Response) => {
-      res.redirect(307, '/oauth/register');
-    });
 
     // OAuth 2.0 routes for ChatGPT integration (public)
     this.app.use('/oauth', createOAuthRouter(this.app_.oauthService));
